Added function to calculate closeness centrality.
This function uses the jgrapht-sna implementation of Freeman closeness centrality, which uses the Floyd-Warshall algorithm to calculate all possible shortest paths. The calculations are intense and can take a long time to complete. For example, the calculation on Nantes took about 1 hour and 45 minutes.

For the moment, only Freeman closeness centrality with all weights equal to 1 has been implemented.