Make orientation optional; finish table fn conv.
I missed a few things when I tried to convert this to a table function:
namely changing the function signature and updating the SQL usage and
javadoc.

In this commit I also make orientation optional and use a
GraphFunctionParser.

I'm currently getting an ArrayOutOfBoundsException because the
ConnectivityInspector is looking for a weight which these graphs no
longer have. (They are weighted graphs, but I removed the calls to
setWeightField().) The problem is that the current version of
GDMS-Topology is a mix between JGraphT and GraphHopper, and the current
graph creators create GraphHopper graphs. My options are:
- Find an easy way to pass unweighted JGraphT graphs to the connectivity
  inspector, or
- Write my own connectivity inspector in GraphHopper-SNA and use that
  instead.