Modify graph creators to load edges properly.
Undirected graphs (pseudographs or weighted pseudographs) load edges
only once since JGraphT is smart enough to know that they are
bidirectional.

In directed and reversed graphs (directed pseudographs or directed
weighted pseudographs) it is now possible to specify individual edge
orientations (directed, reversed or undirected).

It remains to finish writing all the tests to make sure this works as
expected. Once that is finished, #27 will be done.